RC-T39 PR#127 fixes
