+++ src/components/ui/gluestack-ui-provider/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { OverlayProvider } from '@gluestack-ui/core/overlay/creator';
 import { ToastProvider } from '@gluestack-ui/core/toast/creator';
-import React, { useEffect } from 'react';
+import React, { useLayoutEffect } from 'react';
 import { Appearance, View, type ViewProps } from 'react-native';
 
 export type ModeType = 'light' | 'dark' | 'system';
@@ -9,7 +9,7 @@ export type ModeType = 'light' | 'dark' | 'system';
 export function GluestackUIProvider({ mode = 'light', ...props }: { mode?: ModeType; children?: React.ReactNode; style?: ViewProps['style'] }) {
   // Both the tokens (--color-*) and the `dark:` variant flip through the prefers-color-scheme media
   // query, which react-native-css drives from Appearance — hence the override below.
-  useEffect(() => {
+  useLayoutEffect(() => {
     Appearance.setColorScheme(mode === 'system' ? null : mode);
   }, [mode]);
